requirements.txt
google-generativeai>=0.8.0
pymupdf>=1.24.0
fpdf2>=2.8.0
python-docx>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
rich>=13.7.0
playwright>=1.40.0
nest-asyncio>=1.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
groq>=0.11.0
duckduckgo-search>=6.0.0
